+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 
 ### VB -> C#
+* Fix lost event subscription for `Handles` clause on a sub-property of a `WithEvents` member, e.g. `Handles TextEdit1.Properties.Click` - [#1273](https://github.com/icsharpcode/CodeConverter/issues/1273)
+* Fix name qualification being skipped for a whole file containing such a `Handles` clause - [#1273](https://github.com/icsharpcode/CodeConverter/issues/1273)
 
 
 ### C# -> VB